Project 3-d compositional data onto x-y plane

Points in the 3-d simplex space with coordinates (x, y ,z) such that x + y + z = 1 are projected into the 2-d plane they reside in. This function can be used to convert the 3-d compositional data into 2-d and then be overlayed on the plots output by ternary_plot, conditional_ternary_plot and grouped_ternary_plot.

Usage

prop_to_tern_proj(data, prop, x = ".x", y = ".y")

tern_to_prop_proj(data, x, y, prop = c("p1", "p2", "p3"))

Arguments

data

A data-frame containing the x-y coordinates of the points.

prop

A character vector specifying the columns names of variable containing the projected compositions. Default is "p1", "p2", and "p3".

x

A character string specifying the name for the column containing the x component of the x-y projection of the simplex.

y

A character string specifying the name for the column containing the y component of the x-y projection of the simplex.

Value

A data-frame with the following two columns appended

.x (or value specified in "x")

The x component of the x-y projection of the simplex point.

.y (or value specified in "y")

The y component of the x-y projection of the simplex point.

A data-frame with the following three columns appended

p1 (or first value specified in "prop")

The first component of the 3-d simplex point.

p2 (or second value specified in "prop")

The second component of the 3-d simplex point.

p3 (or third value specified in "prop")

The third component of the 3-d simplex point.

Examples

library(DImodels)
data(sim0)

prop_to_tern_proj(data = sim0, prop = c("p1", "p2", "p3"))
#>           p1        p2        p3   .x        .y community richness response
#> 1  1.0000000 0.0000000 0.0000000 0.50 0.8660254         1        1   24.855
#> 2  0.0000000 1.0000000 0.0000000 0.00 0.0000000         2        1   19.049
#> 3  0.0000000 0.0000000 1.0000000 1.00 0.0000000         3        1   16.292
#> 4  0.8000000 0.2000000 0.0000000 0.40 0.6928203         4        2   31.529
#> 5  0.2000000 0.8000000 0.0000000 0.10 0.1732051         5        2   25.102
#> 6  0.8000000 0.0000000 0.2000000 0.60 0.6928203         6        2   24.615
#> 7  0.2000000 0.0000000 0.8000000 0.90 0.1732051         7        2   18.654
#> 8  0.0000000 0.8000000 0.2000000 0.20 0.0000000         8        2   24.697
#> 9  0.0000000 0.2000000 0.8000000 0.80 0.0000000         9        2   25.017
#> 10 0.5000000 0.5000000 0.0000000 0.25 0.4330127        10        2   32.743
#> 11 0.5000000 0.0000000 0.5000000 0.75 0.4330127        11        2   25.320
#> 12 0.0000000 0.5000000 0.5000000 0.50 0.0000000        12        2   30.214
#> 13 0.6000000 0.2000000 0.2000000 0.50 0.5196152        13        3   30.511
#> 14 0.2000000 0.6000000 0.2000000 0.30 0.1732051        14        3   31.005
#> 15 0.2000000 0.2000000 0.6000000 0.70 0.1732051        15        3   28.964
#> 16 0.3333333 0.3333333 0.3333333 0.50 0.2886751        16        3   32.091
#> 17 1.0000000 0.0000000 0.0000000 0.50 0.8660254         1        1   23.754
#> 18 0.0000000 1.0000000 0.0000000 0.00 0.0000000         2        1   17.712
#> 19 0.0000000 0.0000000 1.0000000 1.00 0.0000000         3        1   15.059
#> 20 0.8000000 0.2000000 0.0000000 0.40 0.6928203         4        2   27.934
#> 21 0.2000000 0.8000000 0.0000000 0.10 0.1732051         5        2   25.150
#> 22 0.8000000 0.0000000 0.2000000 0.60 0.6928203         6        2   24.684
#> 23 0.2000000 0.0000000 0.8000000 0.90 0.1732051         7        2   19.160
#> 24 0.0000000 0.8000000 0.2000000 0.20 0.0000000         8        2   24.167
#> 25 0.0000000 0.2000000 0.8000000 0.80 0.0000000         9        2   21.433
#> 26 0.5000000 0.5000000 0.0000000 0.25 0.4330127        10        2   28.950
#> 27 0.5000000 0.0000000 0.5000000 0.75 0.4330127        11        2   28.736
#> 28 0.0000000 0.5000000 0.5000000 0.50 0.0000000        12        2   28.484
#> 29 0.6000000 0.2000000 0.2000000 0.50 0.5196152        13        3   30.371
#> 30 0.2000000 0.6000000 0.2000000 0.30 0.1732051        14        3   32.420
#> 31 0.2000000 0.2000000 0.6000000 0.70 0.1732051        15        3   26.795
#> 32 0.3333333 0.3333333 0.3333333 0.50 0.2886751        16        3   31.554
#> 33 1.0000000 0.0000000 0.0000000 0.50 0.8660254         1        1   26.285
#> 34 0.0000000 1.0000000 0.0000000 0.00 0.0000000         2        1   18.600
#> 35 0.0000000 0.0000000 1.0000000 1.00 0.0000000         3        1   11.824
#> 36 0.8000000 0.2000000 0.0000000 0.40 0.6928203         4        2   29.437
#> 37 0.2000000 0.8000000 0.0000000 0.10 0.1732051         5        2   25.430
#> 38 0.8000000 0.0000000 0.2000000 0.60 0.6928203         6        2   27.198
#> 39 0.2000000 0.0000000 0.8000000 0.90 0.1732051         7        2   25.959
#> 40 0.0000000 0.8000000 0.2000000 0.20 0.0000000         8        2   29.477
#> 41 0.0000000 0.2000000 0.8000000 0.80 0.0000000         9        2   22.065
#> 42 0.5000000 0.5000000 0.0000000 0.25 0.4330127        10        2   29.250
#> 43 0.5000000 0.0000000 0.5000000 0.75 0.4330127        11        2   26.559
#> 44 0.0000000 0.5000000 0.5000000 0.50 0.0000000        12        2   27.891
#> 45 0.6000000 0.2000000 0.2000000 0.50 0.5196152        13        3   31.985
#> 46 0.2000000 0.6000000 0.2000000 0.30 0.1732051        14        3   27.489
#> 47 0.2000000 0.2000000 0.6000000 0.70 0.1732051        15        3   25.830
#> 48 0.3333333 0.3333333 0.3333333 0.50 0.2886751        16        3   32.649
#> 49 1.0000000 0.0000000 0.0000000 0.50 0.8660254         1        1   21.347
#> 50 0.0000000 1.0000000 0.0000000 0.00 0.0000000         2        1   18.015
#> 51 0.0000000 0.0000000 1.0000000 1.00 0.0000000         3        1   15.684
#> 52 0.8000000 0.2000000 0.0000000 0.40 0.6928203         4        2   28.620
#> 53 0.2000000 0.8000000 0.0000000 0.10 0.1732051         5        2   24.614
#> 54 0.8000000 0.0000000 0.2000000 0.60 0.6928203         6        2   25.292
#> 55 0.2000000 0.0000000 0.8000000 0.90 0.1732051         7        2   20.983
#> 56 0.0000000 0.8000000 0.2000000 0.20 0.0000000         8        2   28.751
#> 57 0.0000000 0.2000000 0.8000000 0.80 0.0000000         9        2   25.679
#> 58 0.5000000 0.5000000 0.0000000 0.25 0.4330127        10        2   27.271
#> 59 0.5000000 0.0000000 0.5000000 0.75 0.4330127        11        2   24.927
#> 60 0.0000000 0.5000000 0.5000000 0.50 0.0000000        12        2   25.712
#> 61 0.6000000 0.2000000 0.2000000 0.50 0.5196152        13        3   33.036
#> 62 0.2000000 0.6000000 0.2000000 0.30 0.1732051        14        3   27.074
#> 63 0.2000000 0.2000000 0.6000000 0.70 0.1732051        15        3   25.937
#> 64 0.3333333 0.3333333 0.3333333 0.50 0.2886751        16        3   29.439

# Change names of the x and y projections
prop_to_tern_proj(data = sim0, prop = c("p1", "p2", "p3"),
                  x = "x-proj", y = "y-proj")
#>           p1        p2        p3 x-proj    y-proj community richness response
#> 1  1.0000000 0.0000000 0.0000000   0.50 0.8660254         1        1   24.855
#> 2  0.0000000 1.0000000 0.0000000   0.00 0.0000000         2        1   19.049
#> 3  0.0000000 0.0000000 1.0000000   1.00 0.0000000         3        1   16.292
#> 4  0.8000000 0.2000000 0.0000000   0.40 0.6928203         4        2   31.529
#> 5  0.2000000 0.8000000 0.0000000   0.10 0.1732051         5        2   25.102
#> 6  0.8000000 0.0000000 0.2000000   0.60 0.6928203         6        2   24.615
#> 7  0.2000000 0.0000000 0.8000000   0.90 0.1732051         7        2   18.654
#> 8  0.0000000 0.8000000 0.2000000   0.20 0.0000000         8        2   24.697
#> 9  0.0000000 0.2000000 0.8000000   0.80 0.0000000         9        2   25.017
#> 10 0.5000000 0.5000000 0.0000000   0.25 0.4330127        10        2   32.743
#> 11 0.5000000 0.0000000 0.5000000   0.75 0.4330127        11        2   25.320
#> 12 0.0000000 0.5000000 0.5000000   0.50 0.0000000        12        2   30.214
#> 13 0.6000000 0.2000000 0.2000000   0.50 0.5196152        13        3   30.511
#> 14 0.2000000 0.6000000 0.2000000   0.30 0.1732051        14        3   31.005
#> 15 0.2000000 0.2000000 0.6000000   0.70 0.1732051        15        3   28.964
#> 16 0.3333333 0.3333333 0.3333333   0.50 0.2886751        16        3   32.091
#> 17 1.0000000 0.0000000 0.0000000   0.50 0.8660254         1        1   23.754
#> 18 0.0000000 1.0000000 0.0000000   0.00 0.0000000         2        1   17.712
#> 19 0.0000000 0.0000000 1.0000000   1.00 0.0000000         3        1   15.059
#> 20 0.8000000 0.2000000 0.0000000   0.40 0.6928203         4        2   27.934
#> 21 0.2000000 0.8000000 0.0000000   0.10 0.1732051         5        2   25.150
#> 22 0.8000000 0.0000000 0.2000000   0.60 0.6928203         6        2   24.684
#> 23 0.2000000 0.0000000 0.8000000   0.90 0.1732051         7        2   19.160
#> 24 0.0000000 0.8000000 0.2000000   0.20 0.0000000         8        2   24.167
#> 25 0.0000000 0.2000000 0.8000000   0.80 0.0000000         9        2   21.433
#> 26 0.5000000 0.5000000 0.0000000   0.25 0.4330127        10        2   28.950
#> 27 0.5000000 0.0000000 0.5000000   0.75 0.4330127        11        2   28.736
#> 28 0.0000000 0.5000000 0.5000000   0.50 0.0000000        12        2   28.484
#> 29 0.6000000 0.2000000 0.2000000   0.50 0.5196152        13        3   30.371
#> 30 0.2000000 0.6000000 0.2000000   0.30 0.1732051        14        3   32.420
#> 31 0.2000000 0.2000000 0.6000000   0.70 0.1732051        15        3   26.795
#> 32 0.3333333 0.3333333 0.3333333   0.50 0.2886751        16        3   31.554
#> 33 1.0000000 0.0000000 0.0000000   0.50 0.8660254         1        1   26.285
#> 34 0.0000000 1.0000000 0.0000000   0.00 0.0000000         2        1   18.600
#> 35 0.0000000 0.0000000 1.0000000   1.00 0.0000000         3        1   11.824
#> 36 0.8000000 0.2000000 0.0000000   0.40 0.6928203         4        2   29.437
#> 37 0.2000000 0.8000000 0.0000000   0.10 0.1732051         5        2   25.430
#> 38 0.8000000 0.0000000 0.2000000   0.60 0.6928203         6        2   27.198
#> 39 0.2000000 0.0000000 0.8000000   0.90 0.1732051         7        2   25.959
#> 40 0.0000000 0.8000000 0.2000000   0.20 0.0000000         8        2   29.477
#> 41 0.0000000 0.2000000 0.8000000   0.80 0.0000000         9        2   22.065
#> 42 0.5000000 0.5000000 0.0000000   0.25 0.4330127        10        2   29.250
#> 43 0.5000000 0.0000000 0.5000000   0.75 0.4330127        11        2   26.559
#> 44 0.0000000 0.5000000 0.5000000   0.50 0.0000000        12        2   27.891
#> 45 0.6000000 0.2000000 0.2000000   0.50 0.5196152        13        3   31.985
#> 46 0.2000000 0.6000000 0.2000000   0.30 0.1732051        14        3   27.489
#> 47 0.2000000 0.2000000 0.6000000   0.70 0.1732051        15        3   25.830
#> 48 0.3333333 0.3333333 0.3333333   0.50 0.2886751        16        3   32.649
#> 49 1.0000000 0.0000000 0.0000000   0.50 0.8660254         1        1   21.347
#> 50 0.0000000 1.0000000 0.0000000   0.00 0.0000000         2        1   18.015
#> 51 0.0000000 0.0000000 1.0000000   1.00 0.0000000         3        1   15.684
#> 52 0.8000000 0.2000000 0.0000000   0.40 0.6928203         4        2   28.620
#> 53 0.2000000 0.8000000 0.0000000   0.10 0.1732051         5        2   24.614
#> 54 0.8000000 0.0000000 0.2000000   0.60 0.6928203         6        2   25.292
#> 55 0.2000000 0.0000000 0.8000000   0.90 0.1732051         7        2   20.983
#> 56 0.0000000 0.8000000 0.2000000   0.20 0.0000000         8        2   28.751
#> 57 0.0000000 0.2000000 0.8000000   0.80 0.0000000         9        2   25.679
#> 58 0.5000000 0.5000000 0.0000000   0.25 0.4330127        10        2   27.271
#> 59 0.5000000 0.0000000 0.5000000   0.75 0.4330127        11        2   24.927
#> 60 0.0000000 0.5000000 0.5000000   0.50 0.0000000        12        2   25.712
#> 61 0.6000000 0.2000000 0.2000000   0.50 0.5196152        13        3   33.036
#> 62 0.2000000 0.6000000 0.2000000   0.30 0.1732051        14        3   27.074
#> 63 0.2000000 0.2000000 0.6000000   0.70 0.1732051        15        3   25.937
#> 64 0.3333333 0.3333333 0.3333333   0.50 0.2886751        16        3   29.439
library(DImodels)
data(sim0)

proj_data <- prop_to_tern_proj(data = sim0, prop = c("p1", "p2", "p3"))

tern_to_prop_proj(data = proj_data, x = ".x", y = ".y")
#>      .x        .y        p1            p2        p3 community richness response
#> 1  0.50 0.8660254 1.0000000  0.000000e+00 0.0000000         1        1   24.855
#> 2  0.00 0.0000000 0.0000000  1.000000e+00 0.0000000         2        1   19.049
#> 3  1.00 0.0000000 0.0000000  0.000000e+00 1.0000000         3        1   16.292
#> 4  0.40 0.6928203 0.8000000  2.000000e-01 0.0000000         4        2   31.529
#> 5  0.10 0.1732051 0.2000000  8.000000e-01 0.0000000         5        2   25.102
#> 6  0.60 0.6928203 0.8000000 -1.110223e-16 0.2000000         6        2   24.615
#> 7  0.90 0.1732051 0.2000000  0.000000e+00 0.8000000         7        2   18.654
#> 8  0.20 0.0000000 0.0000000  8.000000e-01 0.2000000         8        2   24.697
#> 9  0.80 0.0000000 0.0000000  2.000000e-01 0.8000000         9        2   25.017
#> 10 0.25 0.4330127 0.5000000  5.000000e-01 0.0000000        10        2   32.743
#> 11 0.75 0.4330127 0.5000000  0.000000e+00 0.5000000        11        2   25.320
#> 12 0.50 0.0000000 0.0000000  5.000000e-01 0.5000000        12        2   30.214
#> 13 0.50 0.5196152 0.6000000  2.000000e-01 0.2000000        13        3   30.511
#> 14 0.30 0.1732051 0.2000000  6.000000e-01 0.2000000        14        3   31.005
#> 15 0.70 0.1732051 0.2000000  2.000000e-01 0.6000000        15        3   28.964
#> 16 0.50 0.2886751 0.3333333  3.333333e-01 0.3333333        16        3   32.091
#> 17 0.50 0.8660254 1.0000000  0.000000e+00 0.0000000         1        1   23.754
#> 18 0.00 0.0000000 0.0000000  1.000000e+00 0.0000000         2        1   17.712
#> 19 1.00 0.0000000 0.0000000  0.000000e+00 1.0000000         3        1   15.059
#> 20 0.40 0.6928203 0.8000000  2.000000e-01 0.0000000         4        2   27.934
#> 21 0.10 0.1732051 0.2000000  8.000000e-01 0.0000000         5        2   25.150
#> 22 0.60 0.6928203 0.8000000 -1.110223e-16 0.2000000         6        2   24.684
#> 23 0.90 0.1732051 0.2000000  0.000000e+00 0.8000000         7        2   19.160
#> 24 0.20 0.0000000 0.0000000  8.000000e-01 0.2000000         8        2   24.167
#> 25 0.80 0.0000000 0.0000000  2.000000e-01 0.8000000         9        2   21.433
#> 26 0.25 0.4330127 0.5000000  5.000000e-01 0.0000000        10        2   28.950
#> 27 0.75 0.4330127 0.5000000  0.000000e+00 0.5000000        11        2   28.736
#> 28 0.50 0.0000000 0.0000000  5.000000e-01 0.5000000        12        2   28.484
#> 29 0.50 0.5196152 0.6000000  2.000000e-01 0.2000000        13        3   30.371
#> 30 0.30 0.1732051 0.2000000  6.000000e-01 0.2000000        14        3   32.420
#> 31 0.70 0.1732051 0.2000000  2.000000e-01 0.6000000        15        3   26.795
#> 32 0.50 0.2886751 0.3333333  3.333333e-01 0.3333333        16        3   31.554
#> 33 0.50 0.8660254 1.0000000  0.000000e+00 0.0000000         1        1   26.285
#> 34 0.00 0.0000000 0.0000000  1.000000e+00 0.0000000         2        1   18.600
#> 35 1.00 0.0000000 0.0000000  0.000000e+00 1.0000000         3        1   11.824
#> 36 0.40 0.6928203 0.8000000  2.000000e-01 0.0000000         4        2   29.437
#> 37 0.10 0.1732051 0.2000000  8.000000e-01 0.0000000         5        2   25.430
#> 38 0.60 0.6928203 0.8000000 -1.110223e-16 0.2000000         6        2   27.198
#> 39 0.90 0.1732051 0.2000000  0.000000e+00 0.8000000         7        2   25.959
#> 40 0.20 0.0000000 0.0000000  8.000000e-01 0.2000000         8        2   29.477
#> 41 0.80 0.0000000 0.0000000  2.000000e-01 0.8000000         9        2   22.065
#> 42 0.25 0.4330127 0.5000000  5.000000e-01 0.0000000        10        2   29.250
#> 43 0.75 0.4330127 0.5000000  0.000000e+00 0.5000000        11        2   26.559
#> 44 0.50 0.0000000 0.0000000  5.000000e-01 0.5000000        12        2   27.891
#> 45 0.50 0.5196152 0.6000000  2.000000e-01 0.2000000        13        3   31.985
#> 46 0.30 0.1732051 0.2000000  6.000000e-01 0.2000000        14        3   27.489
#> 47 0.70 0.1732051 0.2000000  2.000000e-01 0.6000000        15        3   25.830
#> 48 0.50 0.2886751 0.3333333  3.333333e-01 0.3333333        16        3   32.649
#> 49 0.50 0.8660254 1.0000000  0.000000e+00 0.0000000         1        1   21.347
#> 50 0.00 0.0000000 0.0000000  1.000000e+00 0.0000000         2        1   18.015
#> 51 1.00 0.0000000 0.0000000  0.000000e+00 1.0000000         3        1   15.684
#> 52 0.40 0.6928203 0.8000000  2.000000e-01 0.0000000         4        2   28.620
#> 53 0.10 0.1732051 0.2000000  8.000000e-01 0.0000000         5        2   24.614
#> 54 0.60 0.6928203 0.8000000 -1.110223e-16 0.2000000         6        2   25.292
#> 55 0.90 0.1732051 0.2000000  0.000000e+00 0.8000000         7        2   20.983
#> 56 0.20 0.0000000 0.0000000  8.000000e-01 0.2000000         8        2   28.751
#> 57 0.80 0.0000000 0.0000000  2.000000e-01 0.8000000         9        2   25.679
#> 58 0.25 0.4330127 0.5000000  5.000000e-01 0.0000000        10        2   27.271
#> 59 0.75 0.4330127 0.5000000  0.000000e+00 0.5000000        11        2   24.927
#> 60 0.50 0.0000000 0.0000000  5.000000e-01 0.5000000        12        2   25.712
#> 61 0.50 0.5196152 0.6000000  2.000000e-01 0.2000000        13        3   33.036
#> 62 0.30 0.1732051 0.2000000  6.000000e-01 0.2000000        14        3   27.074
#> 63 0.70 0.1732051 0.2000000  2.000000e-01 0.6000000        15        3   25.937
#> 64 0.50 0.2886751 0.3333333  3.333333e-01 0.3333333        16        3   29.439

# Change prop names
tern_to_prop_proj(data = proj_data, x = ".x", y = ".y",
                  prop = c("prop1", "prop2", "prop3"))
#>      .x        .y     prop1         prop2     prop3        p1        p2
#> 1  0.50 0.8660254 1.0000000  0.000000e+00 0.0000000 1.0000000 0.0000000
#> 2  0.00 0.0000000 0.0000000  1.000000e+00 0.0000000 0.0000000 1.0000000
#> 3  1.00 0.0000000 0.0000000  0.000000e+00 1.0000000 0.0000000 0.0000000
#> 4  0.40 0.6928203 0.8000000  2.000000e-01 0.0000000 0.8000000 0.2000000
#> 5  0.10 0.1732051 0.2000000  8.000000e-01 0.0000000 0.2000000 0.8000000
#> 6  0.60 0.6928203 0.8000000 -1.110223e-16 0.2000000 0.8000000 0.0000000
#> 7  0.90 0.1732051 0.2000000  0.000000e+00 0.8000000 0.2000000 0.0000000
#> 8  0.20 0.0000000 0.0000000  8.000000e-01 0.2000000 0.0000000 0.8000000
#> 9  0.80 0.0000000 0.0000000  2.000000e-01 0.8000000 0.0000000 0.2000000
#> 10 0.25 0.4330127 0.5000000  5.000000e-01 0.0000000 0.5000000 0.5000000
#> 11 0.75 0.4330127 0.5000000  0.000000e+00 0.5000000 0.5000000 0.0000000
#> 12 0.50 0.0000000 0.0000000  5.000000e-01 0.5000000 0.0000000 0.5000000
#> 13 0.50 0.5196152 0.6000000  2.000000e-01 0.2000000 0.6000000 0.2000000
#> 14 0.30 0.1732051 0.2000000  6.000000e-01 0.2000000 0.2000000 0.6000000
#> 15 0.70 0.1732051 0.2000000  2.000000e-01 0.6000000 0.2000000 0.2000000
#> 16 0.50 0.2886751 0.3333333  3.333333e-01 0.3333333 0.3333333 0.3333333
#> 17 0.50 0.8660254 1.0000000  0.000000e+00 0.0000000 1.0000000 0.0000000
#> 18 0.00 0.0000000 0.0000000  1.000000e+00 0.0000000 0.0000000 1.0000000
#> 19 1.00 0.0000000 0.0000000  0.000000e+00 1.0000000 0.0000000 0.0000000
#> 20 0.40 0.6928203 0.8000000  2.000000e-01 0.0000000 0.8000000 0.2000000
#> 21 0.10 0.1732051 0.2000000  8.000000e-01 0.0000000 0.2000000 0.8000000
#> 22 0.60 0.6928203 0.8000000 -1.110223e-16 0.2000000 0.8000000 0.0000000
#> 23 0.90 0.1732051 0.2000000  0.000000e+00 0.8000000 0.2000000 0.0000000
#> 24 0.20 0.0000000 0.0000000  8.000000e-01 0.2000000 0.0000000 0.8000000
#> 25 0.80 0.0000000 0.0000000  2.000000e-01 0.8000000 0.0000000 0.2000000
#> 26 0.25 0.4330127 0.5000000  5.000000e-01 0.0000000 0.5000000 0.5000000
#> 27 0.75 0.4330127 0.5000000  0.000000e+00 0.5000000 0.5000000 0.0000000
#> 28 0.50 0.0000000 0.0000000  5.000000e-01 0.5000000 0.0000000 0.5000000
#> 29 0.50 0.5196152 0.6000000  2.000000e-01 0.2000000 0.6000000 0.2000000
#> 30 0.30 0.1732051 0.2000000  6.000000e-01 0.2000000 0.2000000 0.6000000
#> 31 0.70 0.1732051 0.2000000  2.000000e-01 0.6000000 0.2000000 0.2000000
#> 32 0.50 0.2886751 0.3333333  3.333333e-01 0.3333333 0.3333333 0.3333333
#> 33 0.50 0.8660254 1.0000000  0.000000e+00 0.0000000 1.0000000 0.0000000
#> 34 0.00 0.0000000 0.0000000  1.000000e+00 0.0000000 0.0000000 1.0000000
#> 35 1.00 0.0000000 0.0000000  0.000000e+00 1.0000000 0.0000000 0.0000000
#> 36 0.40 0.6928203 0.8000000  2.000000e-01 0.0000000 0.8000000 0.2000000
#> 37 0.10 0.1732051 0.2000000  8.000000e-01 0.0000000 0.2000000 0.8000000
#> 38 0.60 0.6928203 0.8000000 -1.110223e-16 0.2000000 0.8000000 0.0000000
#> 39 0.90 0.1732051 0.2000000  0.000000e+00 0.8000000 0.2000000 0.0000000
#> 40 0.20 0.0000000 0.0000000  8.000000e-01 0.2000000 0.0000000 0.8000000
#> 41 0.80 0.0000000 0.0000000  2.000000e-01 0.8000000 0.0000000 0.2000000
#> 42 0.25 0.4330127 0.5000000  5.000000e-01 0.0000000 0.5000000 0.5000000
#> 43 0.75 0.4330127 0.5000000  0.000000e+00 0.5000000 0.5000000 0.0000000
#> 44 0.50 0.0000000 0.0000000  5.000000e-01 0.5000000 0.0000000 0.5000000
#> 45 0.50 0.5196152 0.6000000  2.000000e-01 0.2000000 0.6000000 0.2000000
#> 46 0.30 0.1732051 0.2000000  6.000000e-01 0.2000000 0.2000000 0.6000000
#> 47 0.70 0.1732051 0.2000000  2.000000e-01 0.6000000 0.2000000 0.2000000
#> 48 0.50 0.2886751 0.3333333  3.333333e-01 0.3333333 0.3333333 0.3333333
#> 49 0.50 0.8660254 1.0000000  0.000000e+00 0.0000000 1.0000000 0.0000000
#> 50 0.00 0.0000000 0.0000000  1.000000e+00 0.0000000 0.0000000 1.0000000
#> 51 1.00 0.0000000 0.0000000  0.000000e+00 1.0000000 0.0000000 0.0000000
#> 52 0.40 0.6928203 0.8000000  2.000000e-01 0.0000000 0.8000000 0.2000000
#> 53 0.10 0.1732051 0.2000000  8.000000e-01 0.0000000 0.2000000 0.8000000
#> 54 0.60 0.6928203 0.8000000 -1.110223e-16 0.2000000 0.8000000 0.0000000
#> 55 0.90 0.1732051 0.2000000  0.000000e+00 0.8000000 0.2000000 0.0000000
#> 56 0.20 0.0000000 0.0000000  8.000000e-01 0.2000000 0.0000000 0.8000000
#> 57 0.80 0.0000000 0.0000000  2.000000e-01 0.8000000 0.0000000 0.2000000
#> 58 0.25 0.4330127 0.5000000  5.000000e-01 0.0000000 0.5000000 0.5000000
#> 59 0.75 0.4330127 0.5000000  0.000000e+00 0.5000000 0.5000000 0.0000000
#> 60 0.50 0.0000000 0.0000000  5.000000e-01 0.5000000 0.0000000 0.5000000
#> 61 0.50 0.5196152 0.6000000  2.000000e-01 0.2000000 0.6000000 0.2000000
#> 62 0.30 0.1732051 0.2000000  6.000000e-01 0.2000000 0.2000000 0.6000000
#> 63 0.70 0.1732051 0.2000000  2.000000e-01 0.6000000 0.2000000 0.2000000
#> 64 0.50 0.2886751 0.3333333  3.333333e-01 0.3333333 0.3333333 0.3333333
#>           p3 community richness response
#> 1  0.0000000         1        1   24.855
#> 2  0.0000000         2        1   19.049
#> 3  1.0000000         3        1   16.292
#> 4  0.0000000         4        2   31.529
#> 5  0.0000000         5        2   25.102
#> 6  0.2000000         6        2   24.615
#> 7  0.8000000         7        2   18.654
#> 8  0.2000000         8        2   24.697
#> 9  0.8000000         9        2   25.017
#> 10 0.0000000        10        2   32.743
#> 11 0.5000000        11        2   25.320
#> 12 0.5000000        12        2   30.214
#> 13 0.2000000        13        3   30.511
#> 14 0.2000000        14        3   31.005
#> 15 0.6000000        15        3   28.964
#> 16 0.3333333        16        3   32.091
#> 17 0.0000000         1        1   23.754
#> 18 0.0000000         2        1   17.712
#> 19 1.0000000         3        1   15.059
#> 20 0.0000000         4        2   27.934
#> 21 0.0000000         5        2   25.150
#> 22 0.2000000         6        2   24.684
#> 23 0.8000000         7        2   19.160
#> 24 0.2000000         8        2   24.167
#> 25 0.8000000         9        2   21.433
#> 26 0.0000000        10        2   28.950
#> 27 0.5000000        11        2   28.736
#> 28 0.5000000        12        2   28.484
#> 29 0.2000000        13        3   30.371
#> 30 0.2000000        14        3   32.420
#> 31 0.6000000        15        3   26.795
#> 32 0.3333333        16        3   31.554
#> 33 0.0000000         1        1   26.285
#> 34 0.0000000         2        1   18.600
#> 35 1.0000000         3        1   11.824
#> 36 0.0000000         4        2   29.437
#> 37 0.0000000         5        2   25.430
#> 38 0.2000000         6        2   27.198
#> 39 0.8000000         7        2   25.959
#> 40 0.2000000         8        2   29.477
#> 41 0.8000000         9        2   22.065
#> 42 0.0000000        10        2   29.250
#> 43 0.5000000        11        2   26.559
#> 44 0.5000000        12        2   27.891
#> 45 0.2000000        13        3   31.985
#> 46 0.2000000        14        3   27.489
#> 47 0.6000000        15        3   25.830
#> 48 0.3333333        16        3   32.649
#> 49 0.0000000         1        1   21.347
#> 50 0.0000000         2        1   18.015
#> 51 1.0000000         3        1   15.684
#> 52 0.0000000         4        2   28.620
#> 53 0.0000000         5        2   24.614
#> 54 0.2000000         6        2   25.292
#> 55 0.8000000         7        2   20.983
#> 56 0.2000000         8        2   28.751
#> 57 0.8000000         9        2   25.679
#> 58 0.0000000        10        2   27.271
#> 59 0.5000000        11        2   24.927
#> 60 0.5000000        12        2   25.712
#> 61 0.2000000        13        3   33.036
#> 62 0.2000000        14        3   27.074
#> 63 0.6000000        15        3   25.937
#> 64 0.3333333        16        3   29.439